# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "osl_api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.28.0"
  },
  "paths": {
...
    "/api/v1/ris/calculations": {
      "post": {
        "tags": [
          "ris"
        ],
        "operationId": "calculate_ris",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputeRisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "RIS computed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeRisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
    "/api/v1/ris/formulas/{year}": {
      "get": {
        "tags": [
          "ris"
        ],
        "operationId": "get_formula_by_year",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "Edition year",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RIS edition for this year",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RisFormulaResponse"
                }
              }
            }
          },
          "404": {
            "description": "No edition was published for this year"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ComputeRisRequest": {
        "type": "object",
        "required": [
          "bodyweight",
          "total",
          "gender"
        ],
        "properties": {
          "bodyweight": {
            "type": "string"
          },
          "formula_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "total": {
            "type": "string"
          }
        }
      },
      "ComputeRisResponse": {
        "type": "object",
        "required": [
          "ris_score",
          "formula_year"
        ],
        "properties": {
          "formula_year": {
            "type": "integer",
            "format": "int32"
          },
          "ris_score": {
            "type": "string"
          }
        }
      },
...
      "Gender": {
        "type": "string",
        "description": "Which category an athlete competes in.",
        "enum": [
          "M",
          "F",
          "MX"
        ]
      },
      "GenderConstants": {
        "type": "object",
        "required": [
          "a",
          "k",
          "b",
          "v",
          "q"
        ],
        "properties": {
          "a": {
            "type": "number",
            "format": "double"
          },
          "b": {
            "type": "number",
            "format": "double"
          },
          "k": {
            "type": "number",
            "format": "double"
          },
          "q": {
            "type": "number",
            "format": "double"
          },
          "v": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "RisConstants": {
        "type": "object",
        "required": [
          "men",
          "women"
        ],
        "properties": {
          "men": {
            "$ref": "#/components/schemas/GenderConstants"
          },
          "women": {
            "$ref": "#/components/schemas/GenderConstants"
          }
        }
      },
...
      "RisFormulaResponse": {
        "type": "object",
        "required": [
          "year",
          "is_current",
          "credit",
          "constants"
        ],
        "properties": {
          "constants": {
            "$ref": "#/components/schemas/RisConstants"
          },
          "credit": {
            "type": "string"
          },
          "is_current": {
            "type": "boolean"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}